size change , QR CODE SCANNER TO QR SCANNER
